task-b/app/requirements.txt
flask
gunicorn
google-cloud-bigquery